Split J2OrbitalSoA.h into .h.cpp
